pyproject.toml
[project]
name = "cobalt-agent"
version = "0.1.0"
description = "Project Cobalt: Autonomous AI Chief of Staff & Trading System"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Director", email = "director@cobalt-core.com" }
]

# --- CORE DEPENDENCIES (Moved here under [project]) ---
dependencies = [
    # 1. THE BRAIN
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.1",
    "litellm == 1.81.8",
    "openai>=1.0.0",
    # 2. THE FINANCE
    "pandas>=2.2.0",
    "pandas-ta-classic",
    "ta-lib>=0.4.0",
    "mplfinance>=0.12.0",
    "aiohttp>=3.9.0",
    # 3. CHIEF OF STAFF
    "mattermostdriver>=7.0",
    "google-api-python-client>=2.0",
    "gitpython>=3.1.0",
    "schedule>=1.2.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    # 4. INFRASTRUCTURE
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    # 5. SECURITY
    "pyotp>=2.9.0",
    "qrcode>=7.4.0",
    "bcrypt>=4.0.0",
    "passlib>=1.7.0",
    "ddgs>=9.10.0",
    "rich>=14.3.2",
    "requests>=2.32.5",
    "yfinance>=1.1.0",
    "psycopg[binary]>=3.3.2",
    "pgvector>=0.4.2",
    "apscheduler>=3.11.2",
    "cryptography>=46.0.4",
    "opencv-python>=4.13.0.92",
    "mss>=10.1.0",
    "psycopg2-binary>=2.9.11",
]

# --- DEV DEPENDENCIES (New Standard: [dependency-groups]) ---
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ipykernel>=6.29.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["setuptools>=61.0", "uv.build"]
build-backend = "setuptools.build_meta"
[tool.uv.extra-build-dependencies]
cobalt-agent = ["uv.build"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["data", "logs", "plans", "scripts", "dev_utils"]
python_files = "test_*.py"
markers = ["integration: marks tests that require external systems (like Docker)"]

[tool.setuptools.packages.find]
where = ["src"]
